PS Website 3.0, flask, with THE ALGORITHM
When integrating the algorithm it became much easier to remove Django
than run with it, so I simply ported it to Flask which removes a lot of
cruft. I also forward-ported (most of) the content from the 2011
website, and the non-algorithm data (SS, overrides) now lives in
ps_data.json which is also in git. The website will auto-update for
normal PS events, anything else just requires a quick update to the json
file.

A procfile is included which should allow this to automatically start if
uploaded to heroku.