Override is only used in iterators, this is single
In this case the data from the manual events (if it exists) is passed
into the PSEvent along with the date to initialise, allowing the manual
event data to override everything. Override is only used when joining
the iterators as heapq.merge isn't stable and ordering can't be
guaranteed, so we need to know which is the manual event.

It would be nice to rewrite the iterator merging not to use heapq.merge
so it is aware of which the manual events are during the merge process,
and make the whole thing a little more clear.