Make time_until in PSEvent cope with 0 days
When relativedelta reports 0 days it might mean that it's exactly 1 month until the next event, not that it's only a few hours away because relativedelta rolls day differences of more than a month up into the months attribute (ditto for more months than a year).  To solve this we collapse down any attribute from the relative date that is 0 and then combine the rest as a sentence.

We've also moved the formatting into a global helper in util.py in case it's useful elsewhere.