Bring PubStandards back to life
We have a new, albeit possibly temporary, venue.
This means we now have two different official
venues, from different time periods. Since we
already have the Big Abstraction of `HIATUSES`,
this is implemented as `VENUES` with somewhat
similar thinking: venues serve between particular
dates. Naturally, because we're changing venue at
the end of a hiatus, the dates line up.

Venues are implemented as a namedtuple which is
clearly superior to a regular tuple as used for
hiatuses, although both are creeping towards
wanting to be Proper Classes, so that for instance
it's easier to both understand and test edge cases
of the date range implementations.

Unlike hiatuses, venues have dates that are dates,
not datetimes, which is unlikely to cause
problems unless someone moves London. The lack of
consistency is annoying, but not as annoying as
changing hiatuses, or thinking carefully about how
to actually do things. And FIRST_PUBSTANDARDS,
from the algorithm, is a date not a datetime.